+++ .github/CODEOWNERS
@@ -1011,7 +1011,7 @@
 /packages/devextreme/js/ui/tooltip_types.d.ts                        @DevExpress/devextreme-editors
 
 /packages/devextreme/js/__internal/ui/tooltip/**                     @DevExpress/devextreme-editors
-/packages/devextreme/js/__internal/ui/m_tooltip.ts                   @DevExpress/devextreme-editors
+/packages/devextreme/js/__internal/ui/tooltip.ts                   @DevExpress/devextreme-editors
 
 /packages/devextreme-scss/scss/widgets/base/_tooltip.scss            @DevExpress/devextreme-editors
 /packages/devextreme-scss/scss/widgets/**/tooltip/**                 @DevExpress/devextreme-editors
