+++ packages/devextreme/js/__internal/scheduler/tooltip_strategies/desktop_tooltip_strategy.ts
@@ -4,8 +4,8 @@ import type {
   ContentReadyEvent, InitializedEvent, ItemContextMenuEvent, Properties as ListProperties,
 } from '@js/ui/list';
 import supportUtils from '@ts/core/utils/m_support';
-import Tooltip from '@ts/ui/m_tooltip';
 
+import Tooltip from '../../ui/tooltip';
 import type { AppointmentTooltipItem } from '../types';
 import { TooltipStrategyBase } from './tooltip_strategy_base';
