+++ packages/devextreme/js/__internal/ui/slider/m_slider_tooltip.ts
@@ -3,9 +3,10 @@ import numberLocalization from '@js/common/core/localization/number';
 import $ from '@js/core/renderer';
 import type { Format } from '@js/localization';
 import type { OptionChanged } from '@ts/core/widget/types';
-import type { TooltipProperties } from '@ts/ui/m_tooltip';
-import Tooltip from '@ts/ui/m_tooltip';
 import { SliderTooltipPositionController } from '@ts/ui/slider/slider_tooltip_position_controller';
+
+import type { TooltipProperties } from '../tooltip';
+import Tooltip from '../tooltip';
 
 // NOTE: Visibility is contolled by the 'visible' option
 // and 'dx-slider-tooltip-visible-on-hover' class
